+++ desktop/flipper-common/src/server-types.tsx
@@ -79,6 +79,7 @@ export type DeviceLogLevel =
 export type ConnectionRecordEntry = {
   time: Date;
   type: 'warning' | 'info' | 'error';
+  os: DeviceOS;
   device: string;
   app: string;
   message: string;
@@ -295,6 +296,7 @@ export type FlipperServerCommands = {
   ) => Promise<ClientResponseType>;
   'android-get-emulators': () => Promise<string[]>;
   'android-launch-emulator': (name: string, coldboot: boolean) => Promise<void>;
+  'android-adb-kill': () => Promise<void>;
   'ios-get-simulators': (bootedOnly: boolean) => Promise<IOSDeviceParams[]>;
   'ios-launch-simulator': (udid: string) => Promise<void>;
   'ios-idb-kill': () => Promise<void>;