+++ desktop/flipper-server-core/src/recorder.tsx
@@ -11,6 +11,7 @@ import {
   ClientQuery,
   ConnectionRecordEntry,
   CommandRecordEntry,
+  getLogger,
 } from 'flipper-common';
 import {FlipperServerImpl} from './FlipperServerImpl';
 
@@ -29,20 +30,30 @@ type ConnectionRecorderEvents = {
 };
 
 class Recorder {
-  private flipperServer: FlipperServerImpl | undefined;
+  private flipperServer_: FlipperServerImpl | undefined;
+  private undefinedClientQuery_: ClientQuery = {
+    app: 'NONE',
+    device: 'NONE',
+    medium: 'NONE',
+    os: 'Browser',
+    device_id: '',
+  };
 
   private handler_ = {
     cmd: (payload: CommandEventPayload) => {
-      if (this.flipperServer) {
+      if (this.flipperServer_) {
         const clientQuery = payload.context as ClientQuery | undefined;
 
         const device = clientQuery?.device ?? 'NONE';
         const app = clientQuery?.app ?? 'NONE';
         const medium = clientQuery?.medium ?? 'NONE';
+        const os =
+          clientQuery?.os ?? (payload.cmd.includes('idb') ? 'iOS' : 'Android');
 
         const entry: CommandRecordEntry = {
           time: new Date(),
           type: payload.success ? 'info' : 'error',
+          os,
           device,
           app,
           message: payload.cmd,
@@ -55,9 +66,31 @@ class Recorder {
           troubleshoot: payload.troubleshoot,
         };
 
-        this.flipperServer.emit('connectivity-troubleshoot-cmd', entry);
+        this.flipperServer_.emit('connectivity-troubleshoot-cmd', entry);
       }
     },
+  };
+
+  private log_ = (
+    type: 'info' | 'warning' | 'error',
+    clientQuery: ClientQuery,
+    ...args: any[]
+  ) => {
+    console.log(`[conn][${type}]`, ...args);
+    if (this.flipperServer_) {
+      const entry: ConnectionRecordEntry = {
+        time: new Date(),
+        type,
+        os: clientQuery.os,
+        device: clientQuery.device,
+        app: clientQuery.app,
+        message: args.join(' '),
+        medium: clientQuery.medium,
+      };
+
+      this.flipperServer_.emit('connectivity-troubleshoot-log', entry);
+      getLogger().track('usage', 'connectivity-log', entry);
+    }
   };
 
   event<Event extends keyof ConnectionRecorderEvents>(
@@ -72,29 +105,19 @@ class Recorder {
   }
 
   log(clientQuery: ClientQuery, ...args: any[]) {
-    console.log('[conn]', ...args);
-    if (this.flipperServer) {
-      const entry: ConnectionRecordEntry = {
-        time: new Date(),
-        type: 'info',
-        device: clientQuery.device,
-        app: clientQuery.app,
-        message: args.join(' '),
-        medium: clientQuery.medium,
-      };
+    this.log_('info', clientQuery, args);
+  }
 
-      this.flipperServer.emit('connectivity-troubleshoot-log', entry);
-    }
-  }
   rawError(...args: any[]) {
-    console.error('[conn]', ...args);
+    this.log_('error', this.undefinedClientQuery_, args);
   }
+
   error(clientQuery: ClientQuery, ...args: any[]) {
-    console.error('[conn]', ...args);
+    this.log_('error', clientQuery, args);
   }
 
   enable(flipperServer: FlipperServerImpl) {
-    this.flipperServer = flipperServer;
+    this.flipperServer_ = flipperServer;
   }
 }