+++ desktop/flipper-ui-core/src/reducers/application.tsx
@@ -11,13 +11,6 @@ import {v1 as uuidv1} from 'uuid';
 import {getRenderHostInstance} from 'flipper-frontend-core';
 import {Actions} from './';
 
-export type ToplevelNavigationItem =
-  | 'appinspect'
-  | 'flipperlogs'
-  | 'notification'
-  | 'connectivity'
-  | undefined;
-
 export type LauncherMsg = {
   message: string;
   severity: 'warning' | 'error';
@@ -44,8 +37,8 @@ export type ShareType = {
 } & SubShareType;
 
 export type State = {
-  topLevelSelection: ToplevelNavigationItem;
-  hasLeftSidebar: boolean;
+  isTroubleshootingModalOpen: boolean;
+  isNotificationModalOpen: boolean;
   leftSidebarVisible: boolean;
   rightSidebarVisible: boolean;
   rightSidebarAvailable: boolean;
@@ -59,6 +52,7 @@ export type State = {
 type BooleanActionType =
   | 'hasLeftSidebar'
   | 'leftSidebarVisible'
+  | 'isNotificationModalOpen'
   | 'rightSidebarVisible'
   | 'rightSidebarAvailable';
 
@@ -66,10 +60,6 @@ export type Action =
   | {
       type: BooleanActionType;
       payload?: boolean;
-    }
-  | {
-      type: 'topLevelSelection';
-      payload: ToplevelNavigationItem;
     }
   | {
       type: 'windowIsFocused';
@@ -89,11 +79,16 @@ export type Action =
   | {
       type: 'REMOVE_STATUS_MSG';
       payload: {msg: string; sender: string};
+    }
+  | {
+      type: 'TOGGLE_CONNECTIVITY_MODAL';
     };
 
 export const initialState: () => State = () => ({
   topLevelSelection: 'appinspect',
   hasLeftSidebar: true,
+  isTroubleshootingModalOpen: false,
+  isNotificationModalOpen: false,
   leftSidebarVisible: true,
   rightSidebarVisible: true,
   rightSidebarAvailable: false,
@@ -126,9 +121,9 @@ export default function reducer(
 ): State {
   state = state || initialState();
   if (
-    action.type === 'hasLeftSidebar' ||
     action.type === 'leftSidebarVisible' ||
     action.type === 'rightSidebarVisible' ||
+    action.type === 'isNotificationModalOpen' ||
     action.type === 'rightSidebarAvailable'
   ) {
     const newValue =
@@ -145,18 +140,10 @@ export default function reducer(
         [action.type]: newValue,
       };
     }
-  } else if (action.type === 'topLevelSelection') {
-    const topLevelSelection = action.payload;
-
-    const hasLeftSidebar =
-      topLevelSelection === 'appinspect' ||
-      topLevelSelection === 'notification';
-
+  } else if (action.type === 'TOGGLE_CONNECTIVITY_MODAL') {
     return {
       ...state,
-      leftSidebarVisible: hasLeftSidebar,
-      hasLeftSidebar,
-      topLevelSelection,
+      isTroubleshootingModalOpen: !state.isTroubleshootingModalOpen,
     };
   } else if (action.type === 'windowIsFocused') {
     return {
@@ -200,11 +187,8 @@ export const toggleAction = (
   payload,
 });
 
-export const setTopLevelSelection = (
-  payload: ToplevelNavigationItem,
-): Action => ({
-  type: 'topLevelSelection',
-  payload,
+export const toggleConnectivityModal = (): Action => ({
+  type: 'TOGGLE_CONNECTIVITY_MODAL',
 });
 
 export const toggleLeftSidebarVisible = (payload?: boolean): Action => ({