+++ desktop/flipper-ui-core/src/sandy-chrome/Navbar.tsx
@@ -27,16 +27,14 @@ import {
   BellOutlined,
   BugOutlined,
   LayoutOutlined,
-  MobileOutlined,
   RocketOutlined,
   SettingOutlined,
   WarningOutlined,
 } from '@ant-design/icons';
 import {
-  setTopLevelSelection,
+  toggleConnectivityModal,
   toggleLeftSidebarVisible,
   toggleRightSidebarVisible,
-  ToplevelNavigationItem,
 } from '../reducers/application';
 import PluginManager from '../chrome/plugin-manager/PluginManager';
 import {showEmulatorLauncher} from './appinspect/LaunchEmulator';
@@ -58,11 +56,11 @@ import {
 import UpdateIndicator from '../chrome/UpdateIndicator';
 import {css} from '@emotion/css';
 import constants from '../fb-stubs/constants';
-import {selectPlugin, setStaticView} from '../reducers/connections';
+import {setStaticView} from '../reducers/connections';
 import {StyleGuide} from './StyleGuide';
 import {openDeeplinkDialog} from '../deeplink';
 import SettingsSheet from '../chrome/SettingsSheet';
-import WelcomeScreen, {WelcomeScreenStaticView} from './WelcomeScreen';
+import WelcomeScreen from './WelcomeScreen';
 import {AppSelector} from './appinspect/AppSelector';
 import {
   NavbarScreenRecordButton,
@@ -72,19 +70,9 @@ import {StatusMessage} from './appinspect/AppInspect';
 import {TroubleshootingGuide} from './appinspect/fb-stubs/TroubleshootingGuide';
 import {FlipperDevTools} from '../chrome/FlipperDevTools';
 import {TroubleshootingHub} from '../chrome/TroubleshootingHub';
+import {Notification} from './notification/Notification';
 
 export const Navbar = withTrackingScope(function Navbar() {
-  const {topLevelSelection, selectedDevice, selectedAppId, selectedPlugin} =
-    useStore((state) => {
-      return {
-        topLevelSelection: state.application.topLevelSelection,
-        selectedDevice: state.connections.selectedDevice,
-        selectedAppId: state.connections.selectedAppId,
-        selectedPlugin: state.connections.selectedPlugin,
-      };
-    });
-
-  const dispatch = useDispatch();
   return (
     <Layout.Horizontal
       borderBottom
@@ -98,25 +86,6 @@ export const Navbar = withTrackingScope(function Navbar() {
       }}>
       <Layout.Horizontal style={{gap: 6}}>
         <LeftSidebarToggleButton />
-        <NavbarButton
-          icon={MobileOutlined}
-          label="App Inspect"
-          toggled={topLevelSelection === 'appinspect'}
-          onClick={() => {
-            dispatch(setTopLevelSelection('appinspect'));
-            if (selectedPlugin) {
-              dispatch(
-                selectPlugin({
-                  selectedPlugin,
-                  selectedAppId,
-                  selectedDevice: selectedDevice,
-                }),
-              );
-            } else {
-              dispatch(setStaticView(WelcomeScreenStaticView));
-            }
-          }}
-        />
         <AppSelector />
         <StatusMessage />
         <NavbarScreenshotButton />
@@ -132,7 +101,7 @@ export const Navbar = withTrackingScope(function Navbar() {
       <Layout.Horizontal style={{gap: 6, alignItems: 'center'}}>
         <NoConnectivityWarning />
 
-        <NotificationButton topLevelSelection={topLevelSelection} />
+        <NotificationButton />
         <TroubleshootMenu />
         <ExtrasMenu />
         <RightSidebarToggleButton />
@@ -242,51 +211,57 @@ function ExportEverythingEverywhereAllAtOnceStatusModal({
   );
 }
 
-function NotificationButton({
-  topLevelSelection,
-}: {
-  topLevelSelection: ToplevelNavigationItem;
-}) {
+function NotificationButton() {
+  const store = useStore();
+  const isOpen = useStore((store) => store.application.isNotificationModalOpen);
   const notifications = useStore((state) => state.notifications);
   const activeNotifications = useMemoize(filterNotifications, [
     notifications.activeNotifications,
     notifications.blocklistedPlugins,
     notifications.blocklistedCategories,
   ]);
-  const dispatch = useDispatch();
   return (
-    <NavbarButton
-      icon={BellOutlined}
-      label="Alerts"
-      zIndex={AlertsZIndex}
-      toggled={topLevelSelection === 'notification'}
-      count={activeNotifications.length}
-      onClick={() => dispatch(setTopLevelSelection('notification'))}
-    />
+    <>
+      <NavbarButton
+        icon={BellOutlined}
+        label="Alerts"
+        zIndex={AlertsZIndex}
+        count={activeNotifications.length}
+        onClick={() => {
+          store.dispatch({type: 'isNotificationModalOpen', payload: true});
+        }}
+      />
+      <Modal
+        visible={isOpen}
+        onCancel={() =>
+          store.dispatch({type: 'isNotificationModalOpen', payload: false})
+        }
+        width="70vw"
+        footer={null}>
+        <div style={{minHeight: '80vh', display: 'flex'}}>
+          <Notification />
+        </div>
+      </Modal>
+    </>
   );
 }
 
 function LeftSidebarToggleButton() {
   const dispatch = useDispatch();
-  const hasMainMenu = useStore((state) => state.application.hasLeftSidebar);
   const mainMenuVisible = useStore(
     (state) => state.application.leftSidebarVisible,
   );
 
-  if (hasMainMenu) {
-    return (
-      <NavbarButton
-        label="Sidebar"
-        icon={LayoutOutlined}
-        toggled={mainMenuVisible}
-        onClick={() => {
-          dispatch(toggleLeftSidebarVisible());
-        }}
-      />
-    );
-  }
-
-  return null;
+  return (
+    <NavbarButton
+      label="Sidebar"
+      icon={LayoutOutlined}
+      toggled={mainMenuVisible}
+      onClick={() => {
+        dispatch(toggleLeftSidebarVisible());
+      }}
+    />
+  );
 }
 
 function RightSidebarToggleButton() {
@@ -470,6 +445,8 @@ function TroubleshootMenu() {
   const [status, setStatus] = useState<
     ExportEverythingEverywhereAllAtOnceStatus | undefined
   >();
+  const [isFlipperDevToolsModalOpen, setFlipperDevToolsModalOpen] =
+    useState(false);
 
   const exportEverythingEverywhereAllAtOnceTracked = useTrackedCallback(
     'Debug data export',
@@ -515,8 +492,7 @@ function TroubleshootMenu() {
             <Menu.Item
               key="connectivity"
               onClick={() => {
-                store.dispatch(setTopLevelSelection('connectivity'));
-                store.dispatch(setStaticView(TroubleshootingHub));
+                store.dispatch(toggleConnectivityModal());
               }}>
               Troubleshoot Connectivity
             </Menu.Item>
@@ -530,8 +506,7 @@ function TroubleshootMenu() {
             <Menu.Item
               key="flipperlogs"
               onClick={() => {
-                store.dispatch(setTopLevelSelection('flipperlogs'));
-                store.dispatch(setStaticView(FlipperDevTools));
+                setFlipperDevToolsModalOpen(true);
               }}>
               <Layout.Horizontal center gap="small">
                 Flipper Logs <Badge count={flipperErrorLogCount} />
@@ -548,7 +523,58 @@ function TroubleshootMenu() {
         status={status}
         setStatus={setStatus}
       />
+      <FlipperDevToolsModal
+        isOpen={isFlipperDevToolsModalOpen}
+        onClose={() => setFlipperDevToolsModalOpen(false)}
+      />
+      <TroubleshootingModal />
     </>
+  );
+}
+
+function TroubleshootingModal() {
+  const store = useStore();
+  const isOpen = useStore(
+    (state) => state.application.isTroubleshootingModalOpen,
+  );
+  return (
+    <Modal
+      visible={isOpen}
+      onCancel={() => store.dispatch(toggleConnectivityModal())}
+      width="100%"
+      footer={null}
+      style={{
+        // override default `top: 100px`
+        top: '5vh',
+      }}>
+      <div style={{minHeight: '80vh', width: '100%', display: 'flex'}}>
+        <TroubleshootingHub />
+      </div>
+    </Modal>
+  );
+}
+
+function FlipperDevToolsModal({
+  isOpen,
+  onClose,
+}: {
+  isOpen: boolean;
+  onClose: () => void;
+}) {
+  return (
+    <Modal
+      visible={isOpen}
+      onCancel={onClose}
+      width="100%"
+      footer={null}
+      style={{
+        // override default `top: 100px`
+        top: '5vh',
+      }}>
+      <div style={{minHeight: '85vh', width: '100%', display: 'flex'}}>
+        <FlipperDevTools />
+      </div>
+    </Modal>
   );
 }