+++ desktop/plugins/public/ui-debugger/components/visualizer/Visualization2D.tsx
@@ -9,7 +9,7 @@
 
 import React, {useEffect, useMemo, useRef, useState} from 'react';
 import {Bounds, Coordinate, Id, ClientNode} from '../../ClientTypes';
-import {NestedNode, OnSelectNode} from '../../DesktopTypes';
+import {NestedNode, OnSelectNode, WireFrameMode} from '../../DesktopTypes';
 
 import {
   produce,
@@ -24,6 +24,7 @@ import {head, isEqual, throttle} from 'lodash';
 import {useDelay} from '../../hooks/useDelay';
 import {Tooltip} from 'antd';
 import {TargetModeState, VisualiserControls} from './VisualizerControls';
+import {getNode} from '../../utils/map';
 
 export const Visualization2D: React.FC<
   {
@@ -42,6 +43,7 @@ export const Visualization2D: React.FC<
   const selectedNodeId = useValue(instance.uiState.selectedNode);
   const hoveredNodes = useValue(instance.uiState.hoveredNodes);
   const hoveredNodeId = head(hoveredNodes);
+  const wireFrameMode = useValue(instance.uiState.wireFrameMode);
 
   const [targetMode, setTargetMode] = useState<TargetModeState>({
     state: 'disabled',
@@ -129,6 +131,10 @@ export const Visualization2D: React.FC<
 
   const onClickOverlay = () => {
     instance.uiActions.onSelectNode(hoveredNodeId, 'visualiser');
+    if (hoveredNodeId != null) {
+      instance.uiActions.ensureAncestorsExpanded(hoveredNodeId);
+    }
+
     if (targetMode.state !== 'disabled') {
       setTargetMode({
         state: 'selected',
@@ -141,8 +147,10 @@ export const Visualization2D: React.FC<
   return (
     <Layout.Container>
       <VisualiserControls
+        onSetWireFrameMode={instance.uiActions.onSetWireFrameMode}
+        wireFrameMode={wireFrameMode}
         focusedNode={focusedNodeId}
-        selectedNode={selectedNodeId?.id}
+        selectedNode={getNode(selectedNodeId?.id, nodes)}
         setTargetMode={setTargetMode}
         targetMode={targetMode}
       />
@@ -226,6 +234,9 @@ export const Visualization2D: React.FC<
             />
           )}
           <MemoedVisualizationNode2D
+            wireframeMode={wireFrameMode}
+            isSelectedOrChildOrSelected={false}
+            selectedNode={selectedNodeId?.id}
             node={focusState.focusedRoot}
             onSelectNode={onSelectNode}
           />
@@ -238,19 +249,36 @@ export const Visualization2D: React.FC<
 const MemoedVisualizationNode2D = React.memo(
   Visualization2DNode,
   (prev, next) => {
-    return prev.node === next.node;
+    if (prev.node != next.node || prev.wireframeMode != next.wireframeMode) {
+      return false;
+    }
+    if (next.wireframeMode == 'All') {
+      //if all wire frames are drawn and the root node is the same
+      //then we are safe
+      return true;
+    } else {
+      //with other modes the selected node affects the drawing
+      return prev.selectedNode === next.selectedNode;
+    }
   },
 );
 
 function Visualization2DNode({
+  wireframeMode,
+  isSelectedOrChildOrSelected,
+  selectedNode,
   node,
   onSelectNode,
 }: {
+  wireframeMode: WireFrameMode;
+  isSelectedOrChildOrSelected: boolean;
+  selectedNode?: Id;
   node: NestedNode;
   onSelectNode: OnSelectNode;
 }) {
   const instance = usePlugin(plugin);
 
+  const isSelected = node.id === selectedNode;
   const ref = useRef<HTMLDivElement>(null);
   let nestedChildren: NestedNode[];
 
@@ -267,7 +295,10 @@ function Visualization2DNode({
   }
 
   const children = nestedChildren.map((child) => (
-    <MemoedVisualizationNode2D
+    <Visualization2DNode
+      wireframeMode={wireframeMode}
+      selectedNode={selectedNode}
+      isSelectedOrChildOrSelected={isSelected || isSelectedOrChildOrSelected}
       key={child.id}
       node={child}
       onSelectNode={onSelectNode}
@@ -277,6 +308,11 @@ function Visualization2DNode({
   const isHighlighted = useValue(instance.uiState.highlightedNodes).has(
     node.id,
   );
+
+  const showBorder =
+    wireframeMode === 'All' ||
+    (wireframeMode === 'SelectedAndChildren' && isSelectedOrChildOrSelected) ||
+    (wireframeMode === 'SelectedOnly' && isSelected);
 
   return (
     <div
@@ -293,7 +329,7 @@ function Visualization2DNode({
         opacity: isHighlighted ? 0.3 : 1,
         backgroundColor: isHighlighted ? 'red' : 'transparent',
       }}>
-      <NodeBorder />
+      {showBorder && <NodeBorder />}
 
       {children}
     </div>
@@ -505,18 +541,19 @@ function hitTest(node: NestedNode, mouseCoordinate: Coordinate): NestedNode[] {
       children = [node.children[node.activeChildIdx]];
     }
     const offsetMouseCoord = offsetCoordinate(mouseCoordinate, nodeBounds);
-    let childHit = false;
+    let anyChildHitRecursive = false;
 
     for (const child of children) {
-      childHit = hitTestRec(child, offsetMouseCoord) || childHit;
+      anyChildHitRecursive =
+        hitTestRec(child, offsetMouseCoord) || anyChildHitRecursive;
     }
 
-    const hit = thisNodeHit && !childHit;
+    const hit = thisNodeHit && !anyChildHitRecursive;
     if (hit) {
       res.push(node);
     }
 
-    return hit;
+    return thisNodeHit || anyChildHitRecursive;
   }
 
   hitTestRec(node, mouseCoordinate);